// extension/src/client/models/RangeInFile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Continue API
 * Continue API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Range } from './Range';
import {
    RangeFromJSON,
    RangeFromJSONTyped,
    RangeToJSON,
} from './Range';

/**
 * 
 * @export
 * @interface RangeInFile
 */
export interface RangeInFile {
    /**
     * 
     * @type {string}
     * @memberof RangeInFile
     */
    filepath: string;
    /**
     * 
     * @type {Range}
     * @memberof RangeInFile
     */
    range: Range;
}

/**
 * Check if a given object implements the RangeInFile interface.
 */
export function instanceOfRangeInFile(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "filepath" in value;
    isInstance = isInstance && "range" in value;

    return isInstance;
}

export function RangeInFileFromJSON(json: any): RangeInFile {
    return RangeInFileFromJSONTyped(json, false);
}

export function RangeInFileFromJSONTyped(json: any, ignoreDiscriminator: boolean): RangeInFile {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'filepath': json['filepath'],
        'range': RangeFromJSON(json['range']),
    };
}

export function RangeInFileToJSON(value?: RangeInFile | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'filepath': value.filepath,
        'range': RangeToJSON(value.range),
    };
}